solved MAJOR bug that hanged the character in air && added scrolling and game over if character fell below game

+++ game.h
@@ -15,7 +15,7 @@
 typedef struct{
 
 	int x,y;
-	int dx,dy;
+	float dx,dy;
 	int animFrame;
 	int onLedge;
 	int faceLeft;
